feat!: config file
This commit adds support for a config file supplied in the directory of
the program. There are no default values, so everything needs to be
supplied in the config file.

Future work would involve creating defaults for things unlikely to be
changed (such as command_time) and also making a option to read from a
specific config file via cmdline args.

Additionally, the config file supports multiple thermocouples at this
point, but the program does not so be warned.