feat: file export
This commit adds export to XLSX files. This calculates the linear
regression and gives a intercept and slope for every thermalcouple in
the device.

The program is now 'feature complete' from a command line perspective
and just needs bug-fixes and ui changes to make it better. Namely, the
config file should supply defaults for options which that is acceptable.

Additionally, people have requested a GUI, so making a simple one that
allows input of device names might be desired from their perspective.